Working: user permissions, monitor, devicesPages, UaerEnabled

+++ webscada/api/src/main/java/webscada/api/dao/IAGenericDao.java
@@ -10,7 +10,7 @@ public interface IAGenericDao<T extends AEntity<Long>> {
 
 	T create(T entity);
 
-	T get(long id);
+	T get(int id);
 
 	void update(T entity);
